Major Change to Get Video Function
Integrated the get video list and get video functions into one. Prompts
for timestamps to pull the list of videos, then provides a list menu to
choose from. Uses the dictonary responses from the API to use exact
timestamps as opposed to asking user to copy and paste the timestamps
they want.